+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # UNRELEASED
 
+### Frontend canister
+
+Sets the `ic_env` cookie for all HTML files only if the canister environment changed in the `commit_batch` method.
+
+- Module hash: c156183e6a6f5c3c71ecde5f38a7280b770ff172b4827127ef03b89b764065ba
+- https://github.com/dfinity/sdk/pull/4446
+
 # 0.30.2
 
 ### Improve frontend canister sync logic